requirements.txt
keras==2.11.0
numpy==1.22.4
tensorflow-cpu==2.11.0
